requirements.txt
fastapi==0.116.1
pandas==2.0.0
pydantic==2.11.8
python-dotenv==1.1.1
SQLAlchemy==2.0.25
uvicorn==0.35.0
psycopg2-binary
openpyxl
python-multipart